+++ Informix.DB2/Tests/UserTests/Issue1303Tests/Tests.UserTests.Issue1303Tests.TestBinary(Informix.DB2,True).sql
@@ -31,8 +31,8 @@ INSERT INTO Issue1303
 VALUES
 (
 	1,
-	@Array,
-	@Binary
+	@Array::BYTE,
+	@Binary::BYTE
 )
 
 BeforeExecute
